% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{as.col_spec}
\alias{as.col_spec}
\title{Generate a column specification}
\usage{
as.col_spec(x)
}
\arguments{
\item{x}{Input object}
}
\value{
a list / S3 object representing column specification
}
\description{
This is most useful for generating a specification using the short form
}
\examples{
as.col_spec("cccnnn")
}
\keyword{internal}
