% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minty.R
\name{mint}
\alias{mint}
\title{Create an Rmd template}
\usage{
mint(template = NULL, path = NULL, name = NULL, verbose = TRUE)
}
\arguments{
\item{template}{[Character] denoting name of the template. See details.}

\item{path}{[Character] A directory path with write permissions where the template will be created}

\item{name}{[Character] An optional character denoting name of the template document and it's parent directory. Defaults to same value as "template".}

\item{verbose}{[Logical] indicating if message is printed to console.}
}
\value{
None
}
\description{
Create an Rmd template in a specified directory
}
\details{
Possible options for argument "template" are "nbis-report", "nbis-report-demo", "nbis-slide", "nbis-slide-demo".
}
\examples{
\dontrun{
mint("nbis-report", "~/Desktop")
mint("nbis-report", "~/Desktop", "my-report")
}

}
